network handler atualizado, mais encapsulado.
